backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-jose
passlib
python-multipart
pandas
python-dotenv
